Inline Markdown image helper
