+++ app/src/main/java/org/ole/planet/myplanet/utilities/Markdown.kt
@@ -37,6 +37,7 @@ import io.noties.markwon.image.network.OkHttpNetworkSchemeHandler
 import io.noties.markwon.movement.MovementMethodPlugin
 import org.commonmark.node.Image
 import org.ole.planet.myplanet.R
+import java.util.regex.Pattern
 
 object Markdown {
     private var currentZoomDialog: Dialog? = null
@@ -110,6 +111,27 @@ object Markdown {
         }
 
         dialog.show()
+    }
+
+    fun prependBaseUrlToImages(
+        markdownContent: String?,
+        baseUrl: String,
+        width: Int = 150,
+        height: Int = 100
+    ): String {
+        val pattern = "!\\[.*?]\\((.*?)\\)"
+        val imagePattern = Pattern.compile(pattern)
+        val matcher = markdownContent?.let { imagePattern.matcher(it) }
+            ?: return markdownContent.orEmpty()
+        val result = StringBuffer()
+        while (matcher.find()) {
+            val relativePath = matcher.group(1)
+            val modifiedPath = relativePath?.replaceFirst("resources/".toRegex(), "")
+            val fullUrl = baseUrl + modifiedPath
+            matcher.appendReplacement(result, "<img src=$fullUrl width=$width height=$height/>")
+        }
+        matcher.appendTail(result)
+        return result.toString()
     }
 
     private class CustomLinkMovementMethod : LinkMovementMethod() {
